🔖v2.2 Squashed commit of the following:
commit bc190deadfec69c0b7d8d050dd5ea2655ef1ff56
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Mon Oct 25 22:00:59 2021 +0200

    🔒️ Protect link deletion
    Now only the owner (and an admin) is able to delete a link.

commit 8a5df1a3493b473bc2a5c402ce0f68a33a11a5a4
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Mon Oct 25 21:34:43 2021 +0200

    🚸 Auto login when user is being registered

commit 6e0b2373b3a2e304c412d624a95954e9ab8859a4
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Mon Oct 25 21:32:56 2021 +0200

    📝 Add missing user Authentication routes

commit 1f1ccb858e79a3a60693841d99dd17cbf32cb559
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Mon Oct 25 20:24:17 2021 +0200

    📝 Update docs
    - Better install description
    - Scaffold for api docs (More precise documentation is coming)

commit fb6e7b85d957a04031c92db8697d8069cb94cd54
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Mon Oct 25 20:21:04 2021 +0200

    💄 Addapt styles for delete button when logged in

commit 139ea415c6bcfa5c465b4346a75a017c35a4c5d2
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Mon Oct 25 20:20:10 2021 +0200

    ⚡️?  Allow browser to cache static filles
    Add .env.sample

commit 07302f72b7d8c3c7e54ace73192423de4340875c
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Mon Oct 25 19:55:15 2021 +0200

    ✨ Give users the option to delte their account
    ⬆️ Update vue to version 3.2.20

commit 760fc5fe9da05d47d9d99d5f80421c8a40d4eee0
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Thu Oct 7 20:45:04 2021 +0200

    🩹 Fix update script

commit 97e351b14f687735b8d382d55456bbc1c6e97cd8
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Thu Oct 7 19:55:57 2021 +0200

    🗃️ Update scripts for chnged database schema
    - relevant for updating without losing data

commit 90a33b19a37874114a245945cc852ea7b4329234
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Thu Oct 7 19:54:36 2021 +0200

    ✨User management for admins
    - the initAdmin can not be deleted
    - There is still an error when a user gets deleted while being logged init
    `Error: Failed to deserialize user out of session`
    - users can be promoted to being an admin but also admins (execpt the initAdmin) can be reduced to users

commit f6b01519036d6fb72258f54a20fb612a0112fe98
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Thu Oct 7 18:10:22 2021 +0200

    💄 Update icon alignment

commit 371ef67b3f3a32dd9498463ec8e6b56d027f9309
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Wed Oct 6 21:46:26 2021 +0200

    💄Make link listing for users nicer
    🩹 also stop the user id from being leaked
    🔥 Remove at this point dead css
    💄 and move like 3 lines of css to a different file

commit 529234736403bbed76e224d397149fa1e3578f79
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Wed Oct 6 20:34:28 2021 +0200

    💄 refactor css to make working with it nicer
    - use css var's

commit 457ddc8201be53c02de37d8c764bf1894c1cf711
Author: ueberflieger <ueberflieger.gaming@gmail.com>
Date:   Sat Oct 2 18:47:02 2021 +0200

    🩹 Fix npm scripts and use nocache library

+++ server/api.js
@@ -50,13 +50,20 @@ function sendAllLinks(user) {
   return data || [];
 }
 
-router.delete("/remove/:id", functions.checkAuthenticated, (req, res) => {
-  db.prepare(`DELETE FROM links WHERE id=?`).run(req.params.id);
-  if (req.isAuthenticated() && req.user.role === "admin") {
+router.delete("/remove/:id", functions.checkAuthenticated, async (req, res) => {
+  if (req.user.role === "admin" || req.user.role === "initAdmin") {
+    db.prepare(`DELETE FROM links WHERE id=?`).run(req.params.id);
     res.send(sendAllLinks());
   } else {
-    res.send(sendAllLinks(req.user.id));
+    let link = await db.prepare(`SELECT * FROM links WHERE id=?`).get(req.params.id);
+    if(link.owner === req.user.id) {
+      db.prepare(`DELETE FROM links WHERE id=?`).run(req.params.id);
+      res.send(sendAllLinks(req.user.id));
+    } else {
+      res.sendStatus(401);
+    }
   }
+
 });
 
 router.get("/all", functions.checkAdminAuthenticated, (req, res) => {
